requirements.txt
pandas
boto3
pendulum
apache-airflow-providers-amazon